added footer and navbar to about page. Moved Footer to the molecules directory
