--- src/components/Organisms/AboutPage/About.js
@@ -1,5 +1,0 @@
-export default function About() {
-  return (
-    <div>About</div>
-  )
-}
