+++ src/components/Organisms/AboutPage/AboutContainer.js
@@ -1,0 +1,11 @@
+import NavBar from "../../Molecules/NavBar/NavBar";
+import Footer from "../../Molecules/Footer/Footer";
+
+export default function AboutContainer() {
+    return (
+        <>
+            <NavBar />
+            <Footer />
+        </>
+    );
+}
