+++ src/pages/AboutPage.js
@@ -1,0 +1,7 @@
+import AboutContainer from "../components/Organisms/AboutPage/AboutContainer"
+
+export default function AboutPage() {
+  return (
+    <AboutContainer />
+  )
+}
