+++ src/pages/Homepage.js
@@ -3,7 +3,7 @@ import Testimonials from "../components/Organisms/HomePage/TestimonialSection/Te
 import Services from "../components/Organisms/HomePage/ServicesSection/Services";
 import Contact from "../components/Organisms/HomePage/ContactSection/ContactSection";
 import NavBar from "../components/Molecules/NavBar/NavBar";
-import Footer from "../components/Organisms/HomePage/Footer/Footer";
+import Footer from "../components/Molecules/Footer/Footer";
 
 export default function Homepage() {
   return (
